The app can now read in a JSONresponse and convert that into earthquake objects.

+++ app/src/main/java/com/example/android/quakereport/Earthquake.java
@@ -6,10 +6,10 @@ import java.util.Date;
 public class Earthquake {
 
     private String location;
-    private float mag;
+    private double mag;
     private Date time;
 
-    public Earthquake(String loc, Date t, float m){
+    public Earthquake(String loc, Date t, double m){
         location = loc;
         time = t;
         mag = m;
@@ -19,7 +19,7 @@ public class Earthquake {
         return location;
     }
 
-    public float getMag(){
+    public double getMag(){
         return mag;
     }
